+++ wdevs-gallery-swiper.php
@@ -16,7 +16,7 @@
  * Plugin Name:          Product Gallery Swiper for WooCommerce
  * Plugin URI:           https://products.wijnberg.dev
  * Description:          Enhances WooCommerce product images display on product listing sections, providing a responsive and touch-friendly swiper.
- * Version:              1.5.6
+ * Version:              1.5.7
  * Author:               Wijnberg Developments
  * License:              GPL-2.0+
  * License URI:          http://www.gnu.org/licenses/gpl-2.0.txt
@@ -40,7 +40,7 @@ if ( ! defined( 'WPINC' ) ) {
  * Start at version 1.0.0 and use SemVer - https://semver.org
  * Rename this for your plugin and update it as you release new versions.
  */
-define( 'WDEVS_GALLERY_SWIPER_VERSION', '1.5.6' );
+define( 'WDEVS_GALLERY_SWIPER_VERSION', '1.5.7' );
 
 /**
  * The code that runs during plugin activation.